+++ droplet.tf
@@ -1,8 +1,8 @@
-# Provising resources in DigitalOcean
+# Provisioning resources in DigitalOcean
 # Droplet, reserved IP, DNS record, volume, volume snapshot
 resource "digitalocean_droplet" "this" {
-  name          = "${var.droplet_name}-${var.droplet_region}"
-  image         = data.digitalocean_droplet_snapshot.this.id
+  name          = local.droplet_name
+  image         = local.image_id
   region        = var.droplet_region
   size          = var.droplet_size
   backups       = var.droplet_backups
@@ -11,10 +11,7 @@ resource "digitalocean_droplet" "this" {
   vpc_uuid      = data.digitalocean_vpc.this.id
   tags          = var.droplet_tags
   user_data     = local.user_data
-  ssh_keys = [
-    data.digitalocean_ssh_key.user.id,
-    data.digitalocean_ssh_key.remote_provisioner.id
-  ]
+  ssh_keys      = local.ssh_keys
 }
 
 resource "digitalocean_project_resources" "this" {
@@ -35,7 +32,7 @@ resource "digitalocean_volume" "this" {
   count = var.droplet_volume_size > 0 ? 1 : 0
 
   region                  = digitalocean_droplet.this.region
-  name                    = "${var.droplet_name}-${var.droplet_region}-volume"
+  name                    = "${local.droplet_name}-${var.droplet_region}-volume"
   size                    = var.droplet_volume_size
   initial_filesystem_type = "ext4"
   description             = "Additional volume for ${digitalocean_droplet.this.name}"
@@ -53,6 +50,6 @@ resource "digitalocean_volume_attachment" "this" {
 resource "digitalocean_volume_snapshot" "this" {
   count = var.droplet_volume_size > 0 ? 1 : 0
 
-  name      = "${var.droplet_name}--${var.droplet_region}-volume-snapshot"
+  name      = "${local.droplet_name}--${var.droplet_region}-volume-snapshot"
   volume_id = digitalocean_volume.this[count.index].id
 }